requirements.txt
pytest
requests
jinja2
jinja2-ansible-filters
stomp.py
